add submission promotion failure reason handling and refactor error logic
